refactor(mypage): components 폴더 위치 변경
