+++ components/mypage/achieveDark.ts
@@ -1,0 +1,17 @@
+import achieveDark1 from "../../public/Achievements/Achieve1_dark.png";
+import achieveDark2 from "../../public/Achievements/Achieve2_dark.png";
+import achieveDark3 from "../../public/Achievements/Achieve3_dark.png";
+import achieveDark4 from "../../public/Achievements/Achieve4_dark.png";
+import achieveDark5 from "../../public/Achievements/Achieve5_dark.png";
+import achieveDark6 from "../../public/Achievements/Achieve6_dark.png";
+import achieveDark7 from "../../public/Achievements/Achieve7_dark.png";
+
+export {
+  achieveDark1,
+  achieveDark2,
+  achieveDark3,
+  achieveDark4,
+  achieveDark5,
+  achieveDark6,
+  achieveDark7,
+};
