+++ components/mypage/profile.ts
@@ -1,0 +1,27 @@
+import profile1 from "../../public/Profile/profile1.png";
+import profile2 from "../../public/Profile/profile2.png";
+import profile3 from "../../public/Profile/profile3.png";
+import profile4 from "../../public/Profile/profile4.png";
+import profile5 from "../../public/Profile/profile5.png";
+import profile6 from "../../public/Profile/profile6.png";
+import profile7 from "../../public/Profile/profile7.png";
+import profile8 from "../../public/Profile/profile8.png";
+import profile9 from "../../public/Profile/profile9.png";
+import profile10 from "../../public/Profile/profile10.png";
+import profile11 from "../../public/Profile/profile11.png";
+import profile12 from "../../public/Profile/profile12.png";
+
+export {
+  profile1,
+  profile2,
+  profile3,
+  profile4,
+  profile5,
+  profile6,
+  profile7,
+  profile8,
+  profile9,
+  profile10,
+  profile11,
+  profile12,
+};
