--- pages/mypage/components/achieveLight.ts
@@ -1,17 +1,0 @@
-import achieveLight1 from "../../../public/Achievements/Achieve1_light.png";
-import achieveLight2 from "../../../public/Achievements/Achieve2_light.png";
-import achieveLight3 from "../../../public/Achievements/Achieve3_light.png";
-import achieveLight4 from "../../../public/Achievements/Achieve4_light.png";
-import achieveLight5 from "../../../public/Achievements/Achieve5_light.png";
-import achieveLight6 from "../../../public/Achievements/Achieve6_light.png";
-import achieveLight7 from "../../../public/Achievements/Achieve7_light.png";
-
-export {
-  achieveLight1,
-  achieveLight2,
-  achieveLight3,
-  achieveLight4,
-  achieveLight5,
-  achieveLight6,
-  achieveLight7,
-};
